hello-langchain-python/1.0/requirements.txt
# LangChain related packages
langchain-core==1.2.11
langchain==1.2.10
langchain-community==0.4.1
langchain-ollama==1.0.1
langchain-openai==1.1.9
langgraph==1.0.8

# Utility packages
python-dotenv==1.2.1
psutil==7.2.2
matplotlib==3.10.8
numpy
soundfile
websocket-client
pyaudio

# API clients
openai==2.20.0
httpx==0.28.1
httpx-sse==0.4.3
qianfan==0.4.12.3

# Others
grandalf==0.8     # Required by langgraph
PyJWT==2.11.0      # Required for authentication